Random Room & Room Exit Generation
Next is placing the objects and making hallways

+++ Assets/Scripts/GameEngine/DungeonEngine/DungeonGen/Generators/GeneratorInterface.cs
@@ -5,7 +5,7 @@ namespace Assets.GameEngine.DungeonEngine.DungeonGen.Generators
     public abstract class GeneratorInterface : ScriptableObject
     {
         // This is the methods that every generator needs, that way it can be swapped out within the DungeonGenerator
-        public abstract void GenerateLevel(float xRadius, float yRadius);
+        public abstract void GenerateLevel(int xRadius, int yRadius);
 
     }
 }
